Before Fixing Python_Lists_And_Dictionaries.py
Okay so in section 5.a, I described a problem about current_character being offset, but it turns out that current_character doesn't actually get offset. So when I tried accounting for the offset via "current_character -= 1", it breaks the program because the offset doesn't occur in the first place. So this version of the script is from before I get rid of this whole offset thing.